+++ enterprise/models.py
@@ -2503,8 +2503,6 @@ class EnterpriseCustomerReportingConfiguration(TimeStampedModel):
         (DELIVERY_METHOD_SFTP, DELIVERY_METHOD_SFTP),
     )
 
-    DATA_TYPE_PROGRESS = 'progress'  # Refers to gathering progress data from Vertica (to be deprecated)
-    DATA_TYPE_PROGRESS_V2 = 'progress_v2'  # Refers to gathering progress data from the Analytics Data API
     DATA_TYPE_PROGRESS_V3 = 'progress_v3'  # Refers to gathering progress data from the Analytics Data API
     DATA_TYPE_CATALOG = 'catalog'
     DATA_TYPE_ENGAGEMENT = 'engagement'  # Refers to gathering engagement data from the Analytics Data API
@@ -2512,8 +2510,7 @@ class EnterpriseCustomerReportingConfiguration(TimeStampedModel):
     DATA_TYPE_COMPLETION = 'completion'
     DATA_TYPE_COURSE_STRUCTURE = 'course_structure'
     DATA_TYPE_CHOICES = (
-        (DATA_TYPE_PROGRESS, DATA_TYPE_PROGRESS),
-        (DATA_TYPE_PROGRESS_V2, DATA_TYPE_PROGRESS_V2),
+
         (DATA_TYPE_PROGRESS_V3, DATA_TYPE_PROGRESS_V3),
         (DATA_TYPE_CATALOG, DATA_TYPE_CATALOG),
         (DATA_TYPE_ENGAGEMENT, DATA_TYPE_ENGAGEMENT),
@@ -2582,7 +2579,7 @@ class EnterpriseCustomerReportingConfiguration(TimeStampedModel):
         max_length=20,
         choices=DATA_TYPE_CHOICES,
         blank=False,
-        default=DATA_TYPE_PROGRESS,
+        default=DATA_TYPE_PROGRESS_V3,
         verbose_name=_("Data Type"),
         help_text=_("The type of data this report should contain.")
     )