+++ integrated_channels/cornerstone/models.py
@@ -163,6 +163,21 @@ class CornerstoneEnterpriseCustomerConfiguration(EnterpriseCustomerPluginConfigu
         """
         return 'CSOD'
 
+    @classmethod
+    def get_by_customer_and_subdomain(cls, enterprise_customer, customer_subdomain):
+        """
+        Get a specific config based on customer + subdomain pair, useful for "muddle of moodles"
+        """
+        # real prod data often has a config like `https://edx.csod.com` alongside `https://edx-stg.csod.com`
+        # if we just did a plain `icontains` using `edx` subdomain, we'd get the staging config too
+        # expanding the subdomain into a proper url prefix lets us get a more exact match.
+        # we require these urls be https
+        subdomain_formatted_as_url_prefix = f'https://{customer_subdomain}.'
+        return cls.objects.select_for_update().filter(
+            enterprise_customer=enterprise_customer,
+            cornerstone_base_url__icontains=subdomain_formatted_as_url_prefix,
+        ).first()
+
     def get_content_metadata_transmitter(self):
         """
         Return a ``CornerstoneContentMetadataTransmitter`` instance.