+++ integrated_channels/sap_success_factors/admin/__init__.py
@@ -68,7 +68,6 @@ class SAPSuccessFactorsEnterpriseCustomerConfigurationAdmin(admin.ModelAdmin):
 
     readonly_fields = (
         'has_access_token',
-        'transmission_chunk_size',
     )
 
     raw_id_fields = (